Incomplete - # 542: Swarming/Multidownload http://dev.powerfolder.com/projects/powerfolder/ticket/542
Fixed a few problems with swarming. This hole thing needs to be 
re-thought, the threading issues are killing me.

Problem: Each Download-Thread + the TransferChecker etc. are 
accessing a DownloadManager. Even thought most interaction is
locked, some can't be and some others require a specific 
chain of events that's not easy this way.